--- a/.gitignore
+++ b/.gitignore
@@ -35,3 +35,4 @@ target.tgz
 .k3d/k3s_manifests/local-storage.yaml
 .k3d/k3s_manifests/rolebindings.yaml
 .eslintcache
+.java-version