--- a/benchmarks/src/test/scala/pl/touk/nussknacker/engine/benchmarks/interpreter/ManyParamsInterpreterBenchmark.scala
+++ b/benchmarks/src/test/scala/pl/touk/nussknacker/engine/benchmarks/interpreter/ManyParamsInterpreterBenchmark.scala
@@ -31,14 +31,14 @@ class ManyParamsInterpreterBenchmark {
     (ctx: Context) => setup(ctx, executionContext)
   }
 
-  private val interpreterSync = prepareInterpreter(SynchronousExecutionContext.create())
-  private val interpreterAsync = prepareInterpreter(ExecutionContext.Implicits.global)
+  private val interpreterSyncIO = prepareInterpreter(SynchronousExecutionContext.ctx)
+  private val interpreterAsyncIO = prepareInterpreter(ExecutionContext.Implicits.global)
 
   @Benchmark
   @BenchmarkMode(Array(Mode.Throughput))
   @OutputTimeUnit(TimeUnit.SECONDS)
   def benchmarkSync(): AnyRef = {
-    interpreterSync(Context("")).unsafeRunSync()
+    interpreterSyncIO(Context("")).unsafeRunSync()
   }
 
 
@@ -46,7 +46,7 @@ class ManyParamsInterpreterBenchmark {
   @BenchmarkMode(Array(Mode.Throughput))
   @OutputTimeUnit(TimeUnit.SECONDS)
   def benchmarkAsync(): AnyRef = {
-    interpreterAsync(Context("")).unsafeRunSync()
+    interpreterAsyncIO(Context("")).unsafeRunSync()
   }
 
 }