--- a/benchmarks/src/test/scala/pl/touk/nussknacker/engine/benchmarks/interpreter/ManyParamsInterpreterBenchmark.scala
+++ b/benchmarks/src/test/scala/pl/touk/nussknacker/engine/benchmarks/interpreter/ManyParamsInterpreterBenchmark.scala
@@ -26,13 +26,13 @@ class ManyParamsInterpreterBenchmark {
     .enricher("e1", "out", "service", (1 to 20).map(i => s"p$i" -> ("''": Expression)): _*)
     .emptySink("sink", "sink")
 
-  private def prepareInterpreter(executionContext: ExecutionContext) = {
+  private def prepareIoInterpreter(executionContext: ExecutionContext) = {
     val setup = new InterpreterSetup[String].sourceInterpretation[IO](process, Map("service" -> new ManyParamsService(executionContext)), Nil)
     (ctx: Context) => setup(ctx, executionContext)
   }
 
-  private val interpreterSyncIO = prepareInterpreter(SynchronousExecutionContext.ctx)
-  private val interpreterAsyncIO = prepareInterpreter(ExecutionContext.Implicits.global)
+  private val interpreterSyncIO = prepareIoInterpreter(SynchronousExecutionContext.ctx)
+  private val interpreterAsyncIO = prepareIoInterpreter(ExecutionContext.Implicits.global)
 
   @Benchmark
   @BenchmarkMode(Array(Mode.Throughput))