--- a/engine/lite/runtime/src/test/scala/pl/touk/nussknacker/engine/lite/sample.scala
+++ b/engine/lite/runtime/src/test/scala/pl/touk/nussknacker/engine/lite/sample.scala
@@ -50,8 +50,8 @@ object sample {
 
     override def monad: Monad[StateType] = implicitly[Monad[StateType]]
 
-    override def fromFuture[T](implicit ec: ExecutionContext): Future[T] => StateType[Either[T, Throwable]] =
-      f => StateT.pure(Await.result(transform(f)(ec), 1 second))
+    override def fromFuture[T]: Future[T] => StateType[Either[T, Throwable]] =
+      f => StateT.pure(Await.result(transform(f), 1 second))
 
   }