--- a/utils/lite-components-testkit/src/main/scala/pl/touk/nussknacker/engine/lite/util/test/SynchronousLiteInterpreter.scala
+++ b/utils/lite-components-testkit/src/main/scala/pl/touk/nussknacker/engine/lite/util/test/SynchronousLiteInterpreter.scala
@@ -1,7 +1,7 @@
 package pl.touk.nussknacker.engine.lite.util.test
 
 import cats.data.{NonEmptyList, Validated}
-import cats.{Id, Monad, catsInstancesForId}
+import cats.{Id, Monad}
 import pl.touk.nussknacker.engine.Interpreter.InterpreterShape
 import pl.touk.nussknacker.engine.Interpreter.InterpreterShape.transform
 import pl.touk.nussknacker.engine.ModelData
@@ -37,7 +37,7 @@ object SynchronousLiteInterpreter {
 
     override def monad: Monad[Id] = Monad[Id]
 
-    override def fromFuture[T](implicit ec: ExecutionContext): Future[T] => Id[Either[T, Throwable]] = f => Await.result(transform(f), waitTime)
+    override def fromFuture[T]: Future[T] => Id[Either[T, Throwable]] = f => Await.result(transform(f), waitTime)
   }
   //todo add generate test data support