Fixed issues:

- `IOShape`: use `IO.apply` instead of `IO.pure` 
- `SyncInterpretationFunction`: in case of IO monad usage, the code should be symmetric to the Future-based implementation (added timeout to the sync run)
- `AsyncInterpretationFunction`: in case of the Future-based implementation - the callback was not invoked
- `interpreter.scala`: converting from Future to Effect (IO or Future) should be done using the sync execution context (now the intention is more clear than before)

IMO the issues above should be solved before bumping cats-effect lib. Ref: https://github.com/TouK/nussknacker/pull/4287

--- 

Benchmarks:

Before changes:
```
[info] Benchmark                                           Mode  Cnt        Score        Error  Units
[info] ManyParamsInterpreterBenchmark.benchmarkAsync      thrpt    8   256944.447 ±  14652.777  ops/s
[info] ManyParamsInterpreterBenchmark.benchmarkSync       thrpt    8   519288.756 ±   4219.996  ops/s
[info] OneParamInterpreterBenchmark.benchmarkAsyncIO      thrpt    8   174094.575 ±   6846.656  ops/s
[info] OneParamInterpreterBenchmark.benchmarkFutureAsync  thrpt    8    93019.418 ±   8049.856  ops/s
[info] OneParamInterpreterBenchmark.benchmarkFutureSync   thrpt    8  2543090.275 ± 135655.268  ops/s
[info] OneParamInterpreterBenchmark.benchmarkSyncIO       thrpt    8  1791769.229 ±  89304.281  ops/s
```

After changes:
```
[info] Benchmark                                           Mode  Cnt        Score       Error  Units
[info] ManyParamsInterpreterBenchmark.benchmarkAsync      thrpt    8   231363.777 ±  1324.431  ops/s
[info] ManyParamsInterpreterBenchmark.benchmarkSync       thrpt    8   422723.163 ± 17596.191  ops/s
[info] OneParamInterpreterBenchmark.benchmarkAsyncIO      thrpt    8   201109.627 ± 23130.016  ops/s
[info] OneParamInterpreterBenchmark.benchmarkFutureAsync  thrpt    8    76491.629 ± 12645.501  ops/s
[info] OneParamInterpreterBenchmark.benchmarkFutureSync   thrpt    8  2571591.618 ± 34357.808  ops/s
[info] OneParamInterpreterBenchmark.benchmarkSyncIO       thrpt    8  1517608.097 ± 33637.059  ops/s
```