@coutoPL 2023-08-23
ref: https://github.com/TouK/nussknacker/pull/4287